functionality on repo add user with permission

+++ util/git.go
@@ -87,17 +87,19 @@ func GenerateRefs(refsPath, repoPath, repoGitName string) {
 
 // UpdateRefsWithNewName ...
 func UpdateRefsWithNewName(refsPath, repoPath, oldRepoGitName, newRepoGitName string) {
-	refsPattern := filepath.Join(refsPath, oldRepoGitName+"*")
+	if oldRepoGitName != newRepoGitName {
+		refsPattern := filepath.Join(refsPath, oldRepoGitName+"*")
 
-	files, err := filepath.Glob(refsPattern)
-	errorhandler.CheckError("Error on updaterefspath filepath.Glob", err)
+		files, err := filepath.Glob(refsPattern)
+		errorhandler.CheckError("Error on updaterefspath filepath.Glob", err)
 
-	for _, f := range files {
-		err := os.Remove(f)
-		errorhandler.CheckError("Error on removing ref files", err)
+		for _, f := range files {
+			err := os.Remove(f)
+			errorhandler.CheckError("Error on removing ref files", err)
+		}
+
+		go GenerateRefs(refsPath, repoPath, newRepoGitName+".git")
 	}
-
-	go GenerateRefs(refsPath, repoPath, newRepoGitName+".git")
 }
 
 // GetCommitCounts ...
